Fix android studio data type warnings

+++ cardscanner-custom/card_corner_detector.h
@@ -21,8 +21,8 @@ public:
     float resizedWidth = 300;        // new width of sized down image
     float detectionAreaRatio = 0.10; // ratio of detection area to image area
     float sigma = 2;                 // higher sigma for more gaussian blur
-    int cannyLowerThreshold = 20;    // reject if pixel gradient below threshold
-    int cannyUpperThreshold = 25;    // accept if pixel gradient above threshold
+    float cannyLowerThreshold = 20;  // reject if pixel gradient below threshold
+    float cannyUpperThreshold = 25;  // accept if pixel gradient above threshold
     int houghlineThreshold = 50;     // minimum intersections to detect a line
     float houghlineMinLineLengthRatio = 0.40; // min length of line to detect
     float houghlineMaxLineGapRatio = 0.20; // max gap between 2 potential lines
